pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graph-foundation-model"
version = "1.0.0"
description = "Graph Foundation Model for Cross-Domain Link Prediction with Few-Shot Meta-Learning"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Graph Foundation Team", email = "team@example.com"}
]
keywords = ["graph-neural-networks", "link-prediction", "meta-learning", "graph-transformer"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "torch>=2.1.0",
    "torch-geometric>=2.4.0",
    "pytorch-lightning>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "networkx>=3.1",
    "matplotlib>=3.7.0",
    "plotly>=5.17.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.4.0",
    "streamlit>=1.28.0",
    "wandb>=0.15.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.9.0",
    "ruff>=0.0.290",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
]

docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/graph-foundation-model"
Documentation = "https://graph-foundation.readthedocs.io"
Repository = "https://github.com/your-org/graph-foundation-model"
Issues = "https://github.com/your-org/graph-foundation-model/issues"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
src = ["config.yaml", "**/*.yaml"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
